requirements.txt
streamlit
pandas
tensorflow
opencv-python-headless
Pillow
pdf2image
streamlit-drawable-canvas
watchdog
